task 2
Task 2.
Pizzeria offers pizza-of-the-day for business lunch. The type of pizza-of-the-day depends on the day of week. Having a pizza-of-the-day simplifies ordering for customers. They don't have to be experts on specific types of pizza. Also, customers can add extra ingredients to the pizza-of-the-day. Write a program that will form orders from customers.